Move to object-store-config naming
