+++ manager/gui/src/components/BreadCrumb.constant.ts
@@ -22,5 +22,5 @@ export const basePathList = [
   "/settings",
   "/maintenance",
   "/configuration",
-  "/users",
+  "/object-store",
 ];
