+++ manager/gui/src/components/object-store-config/LrTenantSecurity.vue
