+++ manager/gui/src/components/object-store-config/LrTenantSetup.vue
