+++ manager/gui/src/scss/style.scss
@@ -68,7 +68,7 @@
   }
 }
 
-.config-page, .users-page{
+.config-page, .object-store-config-page{
   .page-title {
     font-weight: bold;
     font-size: 1.2rem;
